Code used for analyzing survey data
The Framing Effect is a bias that people have when it comes to perceiving information presented in different ways. This effect was first presented by Kahneman and Tversky in 1979 which was based on participant’s decisions to be risk averse or risk seeking in varying scenarios. Since then, the literature on the Framing Effect has expanded greatly to measure phenomena outside of risk. In our study, we wanted to investigate The Attribute Framing Effect which is two equivalent scenarios framed positively or negatively (Peng et al., 2013). We also primed participants with positive or negative faces to prime their affect (Skandrani-Marzouki & Marzouki, 2012). We were curious to know if these conditions had an effect on altruism and/or cognitive effort, so we measured the levels of altruism and cognitive effort through willingness to help and time spent per question. We hypothesized that positively primed participants with positive framing conditions would show higher levels of altruism, and that they would also show higher levels of cognitive effort. We tested ninety-one participants and found that participants given positively framed questions showed higher levels of altruism and willingness to help others. We also found that participants who had been negatively primed with faces and given negatively framed questions showed the highest levels of cognitive effort compared to all other conditions which did not align with our initial hypothesis.