Disease Transmission - Computational Modeling
This is a simple model of infectious disease transmission. Agents will move around a 2-dimensional surface and potentially spread an infectious disease to those nearby.

The model will manipulate the number of agents who move vs. stay in place (i.e. social distance) and observe the effect of social distancing on disease spread.