Conformity Model - Computational Modeling
The purpose of this model is to explore the emergence of fads as a function of varying levels of conformity and anticonformity.

Agents will choose between three options (grow a beard, grow a mustache, or be clean-shaven) based on one of two conformity strategies. Some agents will be "conformists": these agents will adopt facial hair styles in proportion to how popular they are in their social group. Other agents will be "hipsters": these agents will adopt facial hair styles in proportion to how unpopular they are in their social group.